requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic-settings==2.6.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0
boto3==1.35.59
celery[redis]==5.4.0
PyJWT==2.9.0
email-validator==2.2.0
reportlab==4.2.5
svglib==1.5.1
ezdxf==1.3.4
pytest==8.3.3
